# Conflicts:
#	src/main/java/za/ac/cput/entity/Course.java
#	src/main/java/za/ac/cput/entity/CourseModule.java
#	src/main/java/za/ac/cput/factory/CourseFactory.java
#	src/main/java/za/ac/cput/factory/CourseModuleFactory.java
#	src/main/java/za/ac/cput/repository/CourseModuleRepositoryImpl.java
#	src/main/java/za/ac/cput/repository/CourseRepositoryImpl.java
#	src/main/java/za/ac/cput/repository/ICourseModuleRepository.java
#	src/main/java/za/ac/cput/repository/ICourseRepository.java
#	src/test/java/za/ac/cput/factory/CourseFactoryTest.java
#	src/test/java/za/ac/cput/factory/CourseModuleFactoryTest.java
#	src/test/java/za/ac/cput/repository/CourseModuleRepositoryImplTest.java
#	src/test/java/za/ac/cput/repository/CourseRepositoryImplTest.java
#	src/test/java/za/ac/cput/repository/FacultyRepositoryImplTest.java

+++ src/main/java/za/ac/cput/controller/FacultyController.java
@@ -52,7 +52,7 @@ public class FacultyController {
     }
 
     @GetMapping("read/{facultyId}")
-    public ResponseEntity<Faculty> read(@PathVariable int facultyId)
+    public ResponseEntity<Faculty> read(@PathVariable long facultyId)
     {
         Faculty facultyReturned = service.read(facultyId)
         .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
@@ -66,7 +66,7 @@ public class FacultyController {
     }
 
     @DeleteMapping("delete/{facultyId}")
-    public ResponseEntity<Void> deleteById(@PathVariable int facultyId)
+    public ResponseEntity<Void> deleteById(@PathVariable long facultyId)
     {
         service.deleteById(facultyId);
         return ResponseEntity.noContent().build();
