+++ src/main/java/za/ac/cput/factory/FacultyFactory.java
@@ -4,7 +4,7 @@ import za.ac.cput.entity.Faculty;
 
 public class FacultyFactory
 {
-    public static Faculty build(int facultyId, String facultyName, String facultyDescription)
+    public static Faculty build(long facultyId, String facultyName, String facultyDescription)
     {
         return new Faculty.Builder().setFacultyId(facultyId)
                                     .setFacultyName(facultyName)
