+++ src/main/java/za/ac/cput/service/IFacultyService.java
@@ -2,6 +2,6 @@ package za.ac.cput.service;
 
 import za.ac.cput.entity.Faculty;
 
-public interface IFacultyService extends IService<Faculty, Integer>{
-    void deleteById(int facultyId);
+public interface IFacultyService extends IService<Faculty, Long>{
+    void deleteById(Long facultyId);
 }
